[zeroD] Implement Connector base class
