+++ include/cantera/zeroD/FlowDevice.h
@@ -9,6 +9,7 @@
 #include "cantera/base/ct_defs.h"
 #include "cantera/base/global.h"
 #include "cantera/base/ctexceptions.h"
+#include "ConnectorNode.h"
 
 namespace Cantera
 {
@@ -20,33 +21,14 @@ class ReactorBase;
  * connecting reactors.
  * @ingroup flowDeviceGroup
  */
-class FlowDevice
+class FlowDevice : public ConnectorNode
 {
 public:
-    FlowDevice(const string& name="(none)") : m_name(name) {}
+    using ConnectorNode::ConnectorNode;  // inherit constructors
 
-    virtual ~FlowDevice() = default;
-    FlowDevice(const FlowDevice&) = delete;
-    FlowDevice& operator=(const FlowDevice&) = delete;
-
-    //! String indicating the flow device implemented. Usually
-    //! corresponds to the name of the derived class.
-    virtual string type() const {
+    string type() const override {
         return "FlowDevice";
     }
-
-    //! Retrieve flow device name.
-    string name() const {
-        return m_name;
-    }
-
-    //! Set flow device name.
-    void setName(const string& name) {
-        m_name = name;
-    }
-
-    //! Set the default name of a flow device. Returns `false` if it was previously set.
-    bool setDefaultName(map<string, int>& counts);
 
     //! Mass flow rate (kg/s).
     double massFlowRate() {
@@ -133,9 +115,6 @@ public:
     }
 
 protected:
-    string m_name;  //!< Flow device name.
-    bool m_defaultNameSet = false;  //!< `true` if default name has been previously set.
-
     double m_mdot = Undef;
 
     //! Function set by setPressureFunction; used by updateMassFlowRate
